+++ lib/src/core/toast_future.dart
@@ -5,12 +5,15 @@ class ToastFuture {
   final OverlayEntry _entry;
   final VoidCallback _onDismiss;
   bool _isShow = true;
+  Timer timer;
   final GlobalKey<__ToastContainerState> _containerKey;
+  final Duration animationDuration;
 
   ToastFuture._(
     this._entry,
     this._onDismiss,
     this._containerKey,
+    this.animationDuration,
   );
 
   void dismiss({bool showAnim = false}) {
@@ -23,11 +26,14 @@ class ToastFuture {
 
     if (showAnim) {
       _containerKey.currentState.showDismissAnim();
-      Future.delayed(_opacityDuration, () {
+      Future.delayed(animationDuration, () {
         _entry.remove();
       });
     } else {
       _entry.remove();
     }
+
+    timer?.cancel();
+    timer = null;
   }
 }